+++ packages/zowe-explorer-ftp-extension/CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to the "zowe-explorer-ftp-extension" extension will be docum
 - Added support for streams to the `getContents` and `putContents` functions (`FtpMvsApi` and `FtpUssApi` interfaces).
 - **Breaking:** updated the `FtpMvsApi.putContents` function to throw an error when an e-tag conflict is found.
   - This establishes consistency with the `FtpUssApi.putContents` function which has always thrown an error for this scenario.
+- **Breaking:** Removed the deprecated `FtpUssApi.putContents` function in favor of the `FtpUssApi.putContent` function.
+  - The `putContents` function was deprecated in v2 in favor of the replacement function that offers the same capabilities, as well as the feature to upload from a buffer.
 
 ### Bug fixes